+++ gradle-plugin/src/main/kotlin/dev/zacsweers/metro/gradle/MetroPluginExtension.kt
@@ -103,9 +103,9 @@ constructor(layout: ProjectLayout, objects: ObjectFactory, providers: ProviderFa
       )
 
   /**
-   * Enable/disable suggestion to lift @Inject to class if there is no params. Enabled by default.
+   * Enable/disable suggestion to lift @Inject to class when there is only one constructor. Enabled by default.
    */
-  public val suggestClassInjectionIfNoParams: Property<Boolean> =
+  public val warnOnInjectAnnotationPlacement: Property<Boolean> =
     objects.property(Boolean::class.javaObjectType).convention(true)
 
   /**